removed format property moved to subclass.
commented out decode flush as it was throwing exception
